requirements.txt
jupyter
numpy
scipy
matplotlib
scikit-image
opencv-python
pillow
wheel
cython
h5py
tqdm
tensorboard
timm
transforms3d
class-resolver

-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.2+cu113
torchvision==0.11.3+cu113
torchaudio==0.10.2+cu113

-f https://data.pyg.org/whl/torch-1.10.0+cu113.html
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

git+https://github.com/pyg-team/pytorch_geometric.git